+++ novawallet/Common/Model/ChainRegistry/RemoteChain/RemoteChainsModel.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 struct RemoteChainsModel: Codable {
-    let assetDisplayPriorities: [AssetModel.Symbol: UInt16]
-    let chainDisplayPriorities: [ChainModel.Id: UInt16]
+    let assetDisplayPriorities: [AssetModel.Symbol: UInt8]
+    let chainDisplayPriorities: [ChainModel.Id: UInt8]
     let chains: [RemoteChainModel]
 }
